Merges 65 commits from feat/digital-human-backend-20260811, tested at
1029 passed / 0 failed / 2 legitimate skips, into the release branch.

+++ database/migrations/2024_05_22_115717_create_subscription_billing_and_refund_histories_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('subscription_id');
             $table->foreignId('package_id')->nullable();
             $table->enum('transaction_type',['pending_bill','refund'])->default('pending_bill');
-            $table->double('amount', 24, 3);
+            $table->decimal('amount', 24, 3);
             $table->boolean('is_success')->default(0);
             $table->string('reference', 191)->nullable();
             $table->timestamps();
